requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

python-dotenv==1.0.1
requests
openai
yfinance
pandas
numpy
cachetools
pytest
pytest-asyncio
json-repair
PyJWT==2.8.0
cryptography==42.0.5
openai>=1.0.0
requests
aiohttp
